requirements.txt
python-telegram-bot==22.5
anthropic==0.96.0
